+++ apm-agent-plugins/apm-okhttp-plugin/src/test/java/co/elastic/apm/agent/okhttp/OkHttp3ClientAsyncInstrumentationTest.java
@@ -1,0 +1,67 @@
+/*-
+ * #%L
+ * Elastic APM Java agent
+ * %%
+ * Copyright (C) 2018 - 2019 Elastic and contributors
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package co.elastic.apm.agent.okhttp;
+
+import co.elastic.apm.agent.httpclient.AbstractHttpClientInstrumentationTest;
+import okhttp3.Call;
+import okhttp3.Callback;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
+import org.junit.Before;
+
+import java.io.IOException;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+
+public class OkHttp3ClientAsyncInstrumentationTest extends AbstractHttpClientInstrumentationTest {
+
+    private OkHttpClient client;
+
+    @Before
+    public void setUp() {
+        client = new OkHttpClient();
+    }
+
+    @Override
+    protected void performGet(String path) throws Exception {
+        Request request = new Request.Builder()
+            .url(path)
+            .build();
+
+        final CountDownLatch countDownLatch = new CountDownLatch(1);
+        client.newCall(request).enqueue(new Callback() {
+            @Override
+            public void onFailure(Call call, IOException e) {
+                countDownLatch.countDown();
+            }
+
+            @Override
+            public void onResponse(Call call, Response response) {
+                countDownLatch.countDown();
+                response.body().close();
+            }
+        });
+        countDownLatch.await(1, TimeUnit.SECONDS);
+
+    }
+
+}
+